make everything with `*_changes` to be part of the `diff` module.
