+++ crates/but-core/src/diff/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod commit;
+pub use commit::to_commit as commit_to_commit;
+
+mod worktree;
+pub use worktree::status as worktree_status;
