+++ crates/but-core/tests/core/diff/mod.rs
@@ -1,0 +1,2 @@
+mod commit;
+pub(crate) mod worktree;
