+++ crates/but-core/tests/core/main.rs
@@ -1,4 +1,4 @@
 mod commit;
+mod diff;
 mod json_samples;
 mod unified_diff;
-mod worktree;
