+++ crates/but-core/tests/core/unified_diff.rs
@@ -3,7 +3,7 @@ use gix::object::tree::EntryKind;
 
 #[test]
 fn file_added_in_worktree() -> anyhow::Result<()> {
-    let repo = crate::worktree::repo("added-modified-in-worktree")?;
+    let repo = crate::diff::worktree::repo("added-modified-in-worktree")?;
     let actual = extract_patch(UnifiedDiff::compute(
         &repo,
         "modified".into(),
@@ -32,7 +32,7 @@ fn file_added_in_worktree() -> anyhow::Result<()> {
 
 #[test]
 fn binary_text_in_unborn() -> anyhow::Result<()> {
-    let repo = crate::worktree::repo("diff-binary-to-text-unborn")?;
+    let repo = crate::diff::worktree::repo("diff-binary-to-text-unborn")?;
     let actual = extract_patch(UnifiedDiff::compute(
         &repo,
         "file.binary".into(),
@@ -61,7 +61,7 @@ fn binary_text_in_unborn() -> anyhow::Result<()> {
 
 #[test]
 fn binary_text_renamed_unborn() -> anyhow::Result<()> {
-    let repo = crate::worktree::repo("diff-binary-to-text-renamed-in-worktree")?;
+    let repo = crate::diff::worktree::repo("diff-binary-to-text-renamed-in-worktree")?;
     // In case of renames, it uses the name of the previous file for attribute lookups.
     let actual = extract_patch(UnifiedDiff::compute(
         &repo,
@@ -94,7 +94,7 @@ fn binary_text_renamed_unborn() -> anyhow::Result<()> {
 
 #[test]
 fn file_deleted_in_worktree() -> anyhow::Result<()> {
-    let repo = crate::worktree::repo("added-modified-in-worktree")?;
+    let repo = crate::diff::worktree::repo("added-modified-in-worktree")?;
     // Pretending there is no current version does the trick.
     let previous_state = ChangeState {
         id: repo.rev_parse_single("@:modified")?.into(),
@@ -126,7 +126,7 @@ fn file_deleted_in_worktree() -> anyhow::Result<()> {
 
 #[test]
 fn big_file_20_in_worktree() -> anyhow::Result<()> {
-    let mut repo = crate::worktree::repo("big-file-20-unborn")?;
+    let mut repo = crate::diff::worktree::repo("big-file-20-unborn")?;
     repo.config_snapshot_mut()
         .set_value(&gix::config::tree::Core::BIG_FILE_THRESHOLD, "20")?;
     let actual = UnifiedDiff::compute(
@@ -156,7 +156,7 @@ fn big_file_20_in_worktree() -> anyhow::Result<()> {
 
 #[test]
 fn binary_file_in_worktree() -> anyhow::Result<()> {
-    let repo = crate::worktree::repo("binary-file-unborn")?;
+    let repo = crate::diff::worktree::repo("binary-file-unborn")?;
     let actual = UnifiedDiff::compute(
         &repo,
         "with-null-bytes".into(),
@@ -182,7 +182,7 @@ fn binary_file_in_worktree() -> anyhow::Result<()> {
 #[test]
 #[cfg(unix)]
 fn symlink_modified_in_worktree() -> anyhow::Result<()> {
-    let repo = crate::worktree::repo_unix("symlink-change-in-worktree")?;
+    let repo = crate::diff::worktree::repo_unix("symlink-change-in-worktree")?;
     let actual = extract_patch(UnifiedDiff::compute(
         &repo,
         "symlink".into(),
